iasdsa

+++ PrivateCloudFFHS/src/ch/ffhs/privatecloudffhs/ActivityServer.java
@@ -1,22 +1,16 @@
 package ch.ffhs.privatecloudffhs;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Random;
 
 import ch.ffhs.privatecloudffhs.connection.ReadKey;
 import ch.ffhs.privatecloudffhs.connection.RsaKeyGen;
-import ch.ffhs.privatecloudffhs.database.Folder;
 import ch.ffhs.privatecloudffhs.database.PrivateCloudDatabase;
 import ch.ffhs.privatecloudffhs.database.Server;
 import android.app.Activity;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.support.v7.appcompat.R.id;
 import android.util.Log;
 import android.view.View;
-import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.FrameLayout;
@@ -68,7 +62,7 @@ public class ActivityServer extends Activity {
 		
         if(b!=null)
         {
-            serverid =(int) b.getInt("serverid");
+            serverid =b.getInt("serverid");
            if ( serverid == 0)
            {
         	   server = new Server("New Servername");
